+++ changelog.md
@@ -7,6 +7,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 -   [component][typeahead-suggestion] Add keyboard support fot suggestions list
 -   [component][typeahead-suggestion-title] Add highlighting component
 -   [component][typeahead-suggestion] Add initial styles and props
+-   [build] Split icons into a distinct entry
 -   [component][input] Add button control for input
 -   [component][icon] Add component, styles, and icon data
 -   [component][typeahead-search] Add network client